requirements.txt
torch
torchvision
timm
transformers
pandas
pillow
numpy
scikit-learn
opencv-python-headless

